+++ src/main/java/cz/uhk/pro2/tulipani/domain/repository/AppRoleRepository.java
@@ -3,5 +3,5 @@ package cz.uhk.pro2.tulipani.domain.repository;
 import cz.uhk.pro2.tulipani.domain.entity.AppRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppRoleRepository extends JpaRepository<AppRole, Long> {
+public interface AppRoleRepository extends JpaRepository<AppRole, Integer> {
 }
